move Booking Modal to new file
